updated the VidneyPanels to have same functionality as old versions. Moved old classes to subpackage old
