--- health/src/main/java/com/health/gui/GridBagLayoutDemo.java
@@ -1,116 +1,0 @@
-package com.health.gui;
-
-/*
- * Copyright (c) 1995, 2008, Oracle and/or its affiliates. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- *   - Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *
- *   - Redistributions in binary form must reproduce the above copyright
- *     notice, this list of conditions and the following disclaimer in the
- *     documentation and/or other materials provided with the distribution.
- *
- *   - Neither the name of Oracle or the names of its
- *     contributors may be used to endorse or promote products derived
- *     from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
- * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
- * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
- 
-/*
- * GridBagLayoutDemo.java requires no other files.
- */
- 
-import java.awt.Container;
-import java.awt.Dimension;
-import java.awt.GridBagConstraints;
-import java.awt.GridBagLayout;
-
-import javax.swing.JButton;
-import javax.swing.JFrame;
- 
-public class GridBagLayoutDemo {
-    final boolean shouldFill = true;
- 
-    public static void addComponentsToPane(Container pane) { 
-        JButton button;
-    pane.setLayout(new GridBagLayout());
-    GridBagConstraints c = new GridBagConstraints();
-   
-    //natural height, maximum width
-    c.fill = GridBagConstraints.BOTH;
- 
-    button = new JButton("Button 1");
-    c.weightx = 1.0;
-    c.weighty = 1.0;
-    c.gridx = 0;
-    c.gridy = 0;
-    pane.add(button, c);
- 
-    button = new JButton("Button 2");
-    c.weightx = 1.0;
-    c.weighty = 1.0;
-    c.gridx = 1;
-    c.gridy = 0;
-    pane.add(button, c);
- 
-    button = new JButton("Button 3");
-    c.fill = GridBagConstraints.HORIZONTAL;
-    c.weightx = 0.5;
-    c.gridx = 2;
-    c.gridy = 0;
-    pane.add(button, c);
- 
-    button = new JButton("Long-Named Button 4");
-    c.fill = GridBagConstraints.HORIZONTAL;
-    c.ipady = 40;      //make this component tall
-    c.weightx = 0.0;
-    c.gridwidth = 3;
-    c.gridx = 0;
-    c.gridy = 1;
-    pane.add(button, c);
-    }
- 
-    /**
-     * Create the GUI and show it.  For thread safety,
-     * this method should be invoked from the
-     * event-dispatching thread.
-     */
-    private static void createAndShowGUI() {
-        //Create and set up the window.
-        JFrame frame = new JFrame("GridBagLayoutDemo");
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        frame.setPreferredSize(new Dimension(800,600));
-        
-        //Set up the content pane.
-        addComponentsToPane(frame.getContentPane());
- 
-        //Display the window.
-        frame.pack();
-        frame.setVisible(true);
-    }
- 
-    public static void main(String[] args) {
-        //Schedule a job for the event-dispatching thread:
-        //creating and showing this application's GUI.
-        javax.swing.SwingUtilities.invokeLater(new Runnable() {
-            public void run() {
-                createAndShowGUI();
-            }
-        });
-    }
-}
