--- health/src/main/java/com/health/gui/OutputVisualPanel.java
@@ -1,19 +1,0 @@
-package com.health.gui;
-
-import javax.swing.JPanel;
-
-/**
- * class creates and fill the panel for output visuals.
- *
- * @author Daan
- */
-public class OutputVisualPanel extends JPanel {
-
-	/**
-	 * function creates and fill the panel.
-	 */
-	public OutputVisualPanel() {
-
-	}
-
-}
