+++ health/src/main/java/com/health/gui/VInputPanel.java
@@ -3,6 +3,8 @@ package com.health.gui;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 
+import com.health.gui.fileSelection.FileSelectionPanel;
+
 public class VInputPanel extends VidneyPanel {    
     public VInputPanel() {
         super();
