+++ health/src/main/java/com/health/gui/fileSelection/FileSelectionPanel.java
@@ -1,4 +1,4 @@
-package com.health.gui;
+package com.health.gui.fileSelection;
 
 import java.awt.BorderLayout;
 import java.awt.FlowLayout;
