+++ health/src/main/java/com/health/gui/old/OutputDataPanel.java
@@ -1,4 +1,4 @@
-package com.health.gui;
+package com.health.gui.old;
 
 import java.awt.BorderLayout;
