+++ health/src/main/java/com/health/gui/old/ScriptPanel.java
@@ -1,4 +1,4 @@
-package com.health.gui;
+package com.health.gui.old;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
@@ -22,6 +22,9 @@ import org.xml.sax.SAXException;
 
 import com.health.Table;
 import com.health.control.ControlModule;
+import com.health.gui.GUImain;
+import com.health.gui.fileSelection.FileListing;
+import com.health.gui.fileSelection.FileListingRow;
 import com.health.input.Input;
 import com.health.input.InputException;
